refactor: Calculator 함수 LottoResults로 통합
간단한 결과 계산은 LottoResults가 책임지도록 변경
