+++ src/main/java/lotto/util/InputValidator.java
@@ -1,6 +1,4 @@
-package lotto.util.parser;
-
-import static lotto.view.OutputView.printMessage;
+package lotto.util;
 
 import lotto.config.ErrorMessage;
