--- src/main/java/lotto/util/ProfitRateCalculator.java
@@ -1,7 +1,0 @@
-package lotto.util;
-
-public class ProfitRateCalculator {
-    public static double calculate(double totalPrize, double purchaseAmount) {
-        return (totalPrize / purchaseAmount) * 100.0;
-    }
-}
