+++ src/main/java/lotto/util/WinningNumbersParser.java
@@ -1,4 +1,4 @@
-package lotto.util.parser;
+package lotto.util;
 
 import java.util.List;
 import java.util.stream.Collectors;
