+++ src/test/java/lotto/config/LottoRankTest.java
@@ -1,0 +1,4 @@
+package lotto.config;
+
+public class LottoRankTest {
+}
