+++ src/test/java/lotto/domain/entity/BonusNumberTest.java
@@ -1,0 +1,4 @@
+package lotto.domain.entity;
+
+public class BonusNumberTest {
+}
