+++ src/test/java/lotto/domain/entity/LottoResultsTest.java
@@ -1,0 +1,4 @@
+package lotto.domain.entity;
+
+public class LottoResults {
+}
