+++ src/test/java/lotto/domain/entity/LottoTicketsTest.java
@@ -1,0 +1,4 @@
+package lotto.domain.entity;
+
+public class LottoTicketsTest {
+}
