+++ src/test/java/lotto/domain/entity/WinningNumbersTest.java
@@ -1,0 +1,4 @@
+package lotto.domain.entity;
+
+public class WinningNumbersTest {
+}
