+++ src/test/java/lotto/util/RandomNumbersTest.java
@@ -1,0 +1,4 @@
+package lotto.util;
+
+public class RandomNumbersTest {
+}
